Updates to Alexa Adapter to pull out core package with request / activity transform logic, attachment types and context extension methods.
