+++ Bot.Builder.Community.Adapters.Alexa.Core/Attachments/CardAttachment.cs
@@ -1,7 +1,7 @@
 ﻿using Alexa.NET.Response;
 using Microsoft.Bot.Schema;
 
-namespace Bot.Builder.Community.Adapters.Alexa.Attachments
+namespace Bot.Builder.Community.Adapters.Alexa.Core.Attachments
 {
     public class CardAttachment : Attachment
     {
