+++ Bot.Builder.Community.sln
@@ -65,7 +65,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bot.Builder.Community.Adapt
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bot.Builder.Community.Dialogs.Luis", "libraries\Bot.Builder.Community.Dialogs.Luis\Bot.Builder.Community.Dialogs.Luis.csproj", "{2FE9D83B-59AB-4475-BE6F-4803118B1A81}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bot.Builder.Community.Adapters.Alexa.Tests", "tests\Bot.Builder.Community.Adapters.Alexa.Tests\Bot.Builder.Community.Adapters.Alexa.Tests.csproj", "{07ECEE15-6D13-439F-8026-E05B70B749F8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bot.Builder.Community.Adapters.Alexa.Tests", "tests\Bot.Builder.Community.Adapters.Alexa.Tests\Bot.Builder.Community.Adapters.Alexa.Tests.csproj", "{07ECEE15-6D13-439F-8026-E05B70B749F8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bot.Builder.Community.Adapters.Alexa.Core", "Bot.Builder.Community.Adapters.Alexa.Core\Bot.Builder.Community.Adapters.Alexa.Core.csproj", "{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -249,6 +251,14 @@ Global
 		{07ECEE15-6D13-439F-8026-E05B70B749F8}.Documentation|Any CPU.Build.0 = Debug|Any CPU
 		{07ECEE15-6D13-439F-8026-E05B70B749F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{07ECEE15-6D13-439F-8026-E05B70B749F8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858}.Debug - NuGet Packages|Any CPU.ActiveCfg = Debug|Any CPU
+		{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858}.Debug - NuGet Packages|Any CPU.Build.0 = Debug|Any CPU
+		{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858}.Documentation|Any CPU.ActiveCfg = Debug|Any CPU
+		{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858}.Documentation|Any CPU.Build.0 = Debug|Any CPU
+		{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -281,6 +291,7 @@ Global
 		{06CECA83-4578-467E-80DD-66C4D0526B9D} = {BF310E8A-8DA1-441F-90E9-DE0E66553048}
 		{2FE9D83B-59AB-4475-BE6F-4803118B1A81} = {08759E28-8592-4EBA-9A07-19A5BED3FB0C}
 		{07ECEE15-6D13-439F-8026-E05B70B749F8} = {840D4038-9AB8-4750-9FFE-365386CE47E2}
+		{11BB1EAE-FCBC-4E29-AF5E-A813C2F1F858} = {BF310E8A-8DA1-441F-90E9-DE0E66553048}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9FE3B75E-BA2B-45BC-BBF0-DDA8BA10C4F0}
