+++ libraries/Bot.Builder.Community.Adapters.Alexa/Bot.Builder.Community.Adapters.Alexa.csproj
@@ -10,6 +10,7 @@
     <Version>1.0.0</Version>
     <FileVersion>1.0.0</FileVersion>
     <AssemblyVersion>1.0.0</AssemblyVersion>
+    <RepositoryUrl>https://github.com/BotBuilderCommunity/botbuilder-community-dotnet/</RepositoryUrl>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -18,11 +19,15 @@
   </PropertyGroup>
 
   <ItemGroup>
-		<PackageReference Include="Alexa.NET" Version="1.12.0" />
+		<PackageReference Include="Alexa.NET" Version="1.13.0" />
 		<PackageReference Include="Alexa.NET.CustomerProfile" Version="2.1.1" />
 		<PackageReference Include="Microsoft.Bot.Builder" Version="4.7.2" />
 		<PackageReference Include="Microsoft.Bot.Builder.Integration.AspNet.Core" Version="4.7.2" />
 		<PackageReference Include="Microsoft.Net.Http.Headers" Version="2.1.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Bot.Builder.Community.Adapters.Alexa.Core\Bot.Builder.Community.Adapters.Alexa.Core.csproj" />
   </ItemGroup>
 
 </Project>
