--- libraries/Bot.Builder.Community.Adapters.Alexa/Middleware/RequestTransformPatterns.cs
@@ -1,8 +1,0 @@
-﻿namespace Bot.Builder.Community.Adapters.Alexa.Middleware
-{
-    public enum RequestTransformPatterns
-    {
-        MessageActivityTextFromSinglePhraseSlotValue,
-        MessageActivityTextFromIntentAndAllSlotValues
-    }
-}
