+++ libraries/Bot.Builder.Community.Adapters.Alexa/ValidationHelper.cs
@@ -1,15 +1,12 @@
 ﻿using System;
-using System.Linq;
 using System.Threading.Tasks;
 using Alexa.NET.Request;
-using Alexa.NET.Response;
-using Microsoft.AspNetCore.Hosting.Server;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 
 namespace Bot.Builder.Community.Adapters.Alexa
 {
-    internal class AlexaHelper
+    internal class ValidationHelper
     {
         public static async Task<bool> ValidateRequest(HttpRequest request, SkillRequest skillRequest, string body, ILogger logger)
         { 
@@ -44,24 +41,6 @@ namespace Bot.Builder.Community.Adapters.Alexa
             if (!valid || !isTimestampValid)
             {
                 logger.LogError("Validation failed - RequestVerification failed");
-                return false;
-            }
-
-            return true;
-        }
-
-        /// <summary>
-        /// Under certain circumstances, such as the inclusion of certain types of directives
-        /// on a response, should force the 'ShouldEndSession' property not be included on
-        /// an outgoing response. This method determines if this property is allowed to have
-        /// a value assigned.
-        /// </summary>
-        /// <param name="response">Boolean indicating if the 'ShouldEndSession' property can be populated on the response.'</param>
-        /// <returns>bool</returns>
-        internal static bool ShouldSetEndSession(SkillResponse response)
-        {
-            if (response.Response.Directives.Any(d => d is IEndSessionDirective))
-            {
                 return false;
             }
